Reorganize morphir.core to closer align with Elm project
