+++ morphir/core/src/org/finos/morphir/core/errors.scala
@@ -1,2 +1,3 @@
-package org.finos.morphir.core.types
+package org.finos.morphir.core
+
 final case class ParserError(msg: String) extends Exception
