--- morphir/core/src/org/finos/morphir/ir/FQNames.scala
@@ -1,3 +1,0 @@
-package org.finos.morphir.ir
-
-object FQNames {}
