--- morphir/core/src/org/finos/morphir/ir/Names.scala
@@ -1,8 +1,0 @@
-package org.finos.morphir.ir
-import org.finos.morphir.core.types.Naming
-object Names {
-  type Name = Naming.Name
-  val Name = Naming.Name
-
-  // TODO: Fill with the rest of the Name module
-}
