+++ morphir/core/src/org/finos/morphir/ir/Naming.scala
@@ -1,0 +1,13 @@
+package org.finos.morphir.ir
+
+import org.finos.morphir.core.ParserError
+import org.finos.morphir.ir.Package.PackageName
+import org.finos.morphir.ir.Path.Path
+import org.finos.morphir.prelude.Subtype
+
+import scala.annotation.tailrec
+
+object Naming {
+  
+
+}
