--- morphir/core/src/org/finos/morphir/ir/Paths.scala
@@ -1,8 +1,0 @@
-package org.finos.morphir.ir
-import org.finos.morphir.core.types.Naming
-object Paths {
-  type Path = Naming.Path
-  val Path = Naming.Path
-
-  // TODO: Fill in the rest of the Path functions
-}
