+++ morphir/core/src/org/finos/morphir/ir/QName.scala
@@ -1,0 +1,35 @@
+package org.finos.morphir.ir
+
+import Path.Path
+import Name.Name
+import Module.ModulePath
+object QName {
+  final case class QName(modulePath: Path, localName: Name) {
+    @inline def toTuple: (Path, Name) = (modulePath, localName)
+
+    override def toString: String =
+      modulePath.toString(Name.toTitleCase, ".") + ":" + localName.toCamelCase
+
+  }
+
+  object QName {
+    def toTuple(qName: QName): (Path, Name) = qName.toTuple
+
+    def fromTuple(tuple: (Path, Name)): QName = QName(tuple._1, tuple._2)
+
+    def fromName(modulePath: Path, localName: Name): QName = QName(modulePath, localName)
+
+    def getLocalName(qname: QName): Name = qname.localName
+
+    def getModulePath(qname: QName): ModulePath = ModulePath(qname.modulePath)
+
+    def toString(qName: QName): String = qName.toString
+
+    def fromString(str: String): Option[QName] =
+      str.split(":") match {
+        case Array(packageNameString, localNameString) =>
+          Some(QName(Path.fromString(packageNameString), Name.fromString(localNameString)))
+        case _ => None
+      }
+  }
+}
