+++ morphir/core/test/src/org/finos/morphir/core/types/EncodedStringSpec.scala
@@ -2,7 +2,7 @@ package org.finos.morphir.core.types
 
 import org.finos.morphir.testing.MorphirBaseSpec
 import zio.test._
-import org.finos.morphir.core.types.Strings.EncodedString
+import org.finos.morphir.core.Strings.EncodedString
 
 object EncodedStringSpec extends MorphirBaseSpec {
   def spec = suite("EncodedString Spec")(
